Don't override generate_sid
There's no need to override the generate_sid method and call the super
method via an alias because the get_session and destroy_session methods
call generate_sid_with directly. This fixes the issue with the Rails
subclass of Rack::Session::Dalli which overrides the generate_sid method
for compatibility reasons.